Make route line more configurable
