+++ MapboxNavigationUI/MGLMapView.swift
@@ -12,61 +12,6 @@ let arrowLayerIdentifier = "arrowLayer"
 
 extension MGLMapView {
     
-    public func annotate(_ route: Route) {
-        guard var coordinates = route.coordinates else {
-            return
-        }
-        
-        guard let style = style else {
-            return
-        }
-        
-        if let line = style.layer(withIdentifier: routeLayerIdentifier) {
-            style.removeLayer(line)
-        }
-        
-        if let lineCasing = style.layer(withIdentifier: routeLayerCasingIdentifier) {
-            style.removeLayer(lineCasing)
-        }
-        
-        if let source = style.source(withIdentifier: sourceIdentifier) {
-            style.removeSource(source)
-        }
-        
-        let polyline = MGLPolylineFeature(coordinates: &coordinates, count: route.coordinateCount)
-        
-        if let source = style.source(withIdentifier: sourceIdentifier) as? MGLShapeSource {
-            source.shape = polyline
-        } else {
-            let geoJSONSource = MGLShapeSource(identifier: sourceIdentifier, shape: polyline, options: nil)
-            let line = MGLLineStyleLayer(identifier: routeLayerIdentifier, source: geoJSONSource)
-            let lineCasing = MGLLineStyleLayer(identifier: routeLayerCasingIdentifier, source: geoJSONSource)
-            
-            line.lineColor = MGLStyleValue(rawValue: NavigationUI.shared.tintStrokeColor.withAlphaComponent(0.6))
-            line.lineWidth = MGLStyleValue(rawValue: 5)
-            lineCasing.lineColor = MGLStyleValue(rawValue: NavigationUI.shared.tintStrokeColor)
-            lineCasing.lineWidth = MGLStyleValue(rawValue: 9)
-            
-            let cap = NSValue(mglLineCap: .round)
-            let join = NSValue(mglLineJoin: .round)
-            
-            line.lineCap = MGLStyleValue(rawValue: cap)
-            line.lineJoin = MGLStyleValue(rawValue: join)
-            lineCasing.lineCap = MGLStyleValue(rawValue: cap)
-            lineCasing.lineJoin = MGLStyleValue(rawValue: join)
-            
-            style.addSource(geoJSONSource)
-            for layer in style.layers.reversed() {
-                if let layer = layer as? MGLStyleLayer, !(layer is MGLSymbolStyleLayer) &&
-                    layer.identifier != arrowLayerIdentifier && layer.identifier != arrowSourceIdentifier {
-                    style.insertLayer(line, above: layer)
-                    style.insertLayer(lineCasing, below: line)
-                    return
-                }
-            }
-        }
-    }
-    
     public var showsTraffic: Bool {
         get {
             if let style = style {
